Rename poller => pollers
In a future commit, we're going to add a new kind of poller. This commit establishes the 'pollers' package as a collection of poller types. This commit includes a rename of poller.go => system_metrics_poller.go to reflect that.
