+++ src/code.cloudfoundry.org/netmon/pollers/poller_suite_test.go
@@ -1,4 +1,4 @@
-package poller_test
+package pollers_test
 
 import (
 	. "github.com/onsi/ginkgo"
