+++ src/code.cloudfoundry.org/netmon/pollers/system_metrics_poller.go
@@ -1,4 +1,4 @@
-package poller
+package pollers
 
 import (
 	"fmt"
